Remove SideKickPipes, add IPCPipe, fix compile errors. Clean up a bit in usings.
