+++ source/StepBro.ConsoleSidekick.WinForms/MainForm.cs
@@ -1,6 +1,7 @@
 using CommandLine;
 using StepBro.Core.Api;
 using StepBro.Core.Data;
+using StepBro.Core.IPCPipe;
 using StepBro.Core.Logging;
 using StepBro.Core.Tasks;
 using StepBro.Sidekick.Messages;
@@ -19,7 +20,7 @@ namespace StepBro.ConsoleSidekick.WinForms
         private bool m_forceResize = false;
         private bool m_moveToTop = true;
         private bool m_closeRequestedByConsole = false;
-        private SideKickPipe m_pipe = null;
+        private IPCPipe m_pipe = null;
         private Rect m_lastConsolePosition = new Rect();
         private IExecutionAccess m_executingScript = null;
         private PanelsDialog m_panelsDialog = null;
@@ -142,7 +143,7 @@ namespace StepBro.ConsoleSidekick.WinForms
             if (args.Length == 2)
             {
                 m_consoleWindow = nint.Parse(args[1], System.Globalization.NumberStyles.HexNumber);
-                m_pipe = SideKickPipe.StartClient(args[1]);
+                m_pipe = IPCPipe.StartClient("StepBroConsoleSidekick", args[1]);
             }
             else
             {
@@ -1125,11 +1126,11 @@ namespace StepBro.ConsoleSidekick.WinForms
         private class ExecutionAccess : IExecutionAccess, IDisposing
         {
             private MainForm m_parent;
-            private SideKickPipe m_pipe;
+            private IPCPipe m_pipe;
             private TaskExecutionState m_state = TaskExecutionState.StartRequested;
             private bool m_active = true;
 
-            public ExecutionAccess(MainForm parent, SideKickPipe pipe)
+            public ExecutionAccess(MainForm parent, IPCPipe pipe)
             {
                 m_parent = parent;
                 m_pipe = pipe;
