+++ source/stepbro/ConsoleSidekickMessages.cs
@@ -4,10 +4,7 @@ using StepBro.Core.Tasks;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Security.Policy;
-using System.Text;
 using System.Text.Json.Serialization;
-using System.Threading.Tasks;
 
 namespace StepBro.Sidekick.Messages
 {
